// ExistingCode/src/data-access/domain/tables/assayelement/schema.ts
export const AssayElementTableInsertBaseSchema = z.object({
	/**
	 * dataType: uniqueidentifier
	 * defaultValue: (newid())
	 */
	assayElementId: z.any().optional(),
	/**
	 * dataType: nvarchar
	 */
	element: z.string().max(50),
	/**
	 * dataType: nvarchar
	 */
	description: z.string().max(1000).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	elementGroup: z.string().max(50),
	/**
	 * dataType: nvarchar
	 */
	systemUnits: z.string().max(50),
	/**
	 * dataType: smallint
	 * defaultValue: ((0))
	 */
	sortOrder: z.number().int().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	isDefaultInd: z.any().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((1))
	 */
	activeInd: z.any().optional(),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	createdOnDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 * defaultValue: (suser_sname())
	 */
	createdBy: z.string().max(50).optional(),
	/**
	 * dataType: datetimeoffset
	 */
	modifiedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	modifiedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: timestamp
	 */
	rv: z.date(),
});